Possibly Fixed the Meetup Component
